Latest changes.

+++ Adventure/src/adventure/Direction.java
@@ -2,24 +2,24 @@ package adventure;
 
 public enum Direction
 {
-	NOWHERE(0),
-	NORTH(1),
-	SOUTH(2),
-	EAST(3),
-	WEST(4),
-	NORTHEAST(5),
-	NORTHWEST(6),
-	SOUTHEAST(7),
-	SOUTHWEST(8),
-	UP(9),
-	DOWN(10);
+	NOWHERE((byte)(0)),
+	NORTH((byte)(1)),
+	SOUTH((byte)(2)),
+	EAST((byte)(3)),
+	WEST((byte)(4)),
+	NORTHEAST((byte)(5)),
+	NORTHWEST((byte)(6)),
+	SOUTHEAST((byte)(7)),
+	SOUTHWEST((byte)(8)),
+	UP((byte)(9)),
+	DOWN((byte)(10));
 	
 	/*
 	 * Class Instance Variables
 	 * 
 	 */
 	
-	private int	value;
+	private byte value;
 	
 	/*
 	 * Constants
@@ -32,7 +32,7 @@ public enum Direction
 	 * Construtor Method
 	 */
 	
-	private Direction(int direction)
+	private Direction(byte direction)
 	{
 		value = direction;
 	}
