+++ Adventure/src/adventure/LocationTypes.java
@@ -2,28 +2,28 @@ package adventure;
 
 public enum LocationTypes
 {
-	NOWHERE(0),
-	WORLD(1),
-	HOUSE(2),
-	ROOM(3),
-	HALLWAY(4),
-	ENTRANCE(5),
-	DOOR(6),
-	WINDOW(7);
+	NOWHERE((byte)(0)),
+	WORLD((byte)(1)),
+	HOUSE((byte)(2)),
+	ROOM((byte)(3)),
+	HALLWAY((byte)(4)),
+	ENTRANCE((byte)(5)),
+	DOOR((byte)(6)),
+	WINDOW((byte)(7));
 	
 	/*
 	 * Class Instance Variables
 	 * 
 	 */
 	
-	private int	value;
+	private byte value;
 	
 	/*
 	 * Constructor
 	 * 
 	 */
 	
-	private LocationTypes(int type)
+	private LocationTypes(byte type)
 	{
 		value = type;
 	}
